module/dev-requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# module/nmj/test/abstract_test.py: 2
# module/nmj/test/cleaners_test.py: 1
# module/nmj/test/scanner_factory_test.py: 3
pymox == 0.7.8

# module/nmj/scanners/tvdb3.py: 7
pytvdbapi == 0.5.0

# module/nmj/utils.py: 4
requests == 2.18.4

# module/nmj/scanners/tmdb.py: 7
tmdbsimple == 2.7.0

nose
coverage
